requirements.txt
# Minimal Requirements for Liver Fibrosis Staging
# Python 3.12 Compatible
# ===============================================

# Core
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.42.0

# Medical Imaging
pydicom>=2.4.0
SimpleITK>=2.2.0
opencv-python>=4.8.0
scikit-image>=0.21.0
scipy>=1.10.0

# Document Generation
python-docx>=0.8.11

# Utilities
tqdm>=4.66.0
